requirements.txt
gradio==5.34.2
pillow==11.2.1
tensorflow==2.16.2
numpy==1.26.4
faiss-cpu==1.11.0
tqdm==4.67.1
torch==2.7.1
torchvision==0.22.1